Add generation of command line documentation for all modules
