+++ mithril-relay/src/main.rs
@@ -1,79 +1,9 @@
 #![doc = include_str!("../README.md")]
 
 use clap::Parser;
-use config::{builder::DefaultState, ConfigBuilder, Map, Source, Value};
 use mithril_common::StdResult;
-use mithril_relay::RelayCommands;
-use slog::{Drain, Level, Logger};
-use slog_scope::debug;
-use std::{path::PathBuf, sync::Arc};
+use mithril_relay::Args;
 
-#[derive(Parser, Debug, Clone)]
-#[clap(name = "mithril-relay")]
-#[clap(
-    about = "This program is a relay for Mithril nodes.",
-    long_about = None
-)]
-#[command(version)]
-pub struct Args {
-    /// Available commands
-    #[clap(subcommand)]
-    command: RelayCommands,
-
-    /// Run Mode.
-    #[clap(long, env = "RUN_MODE", default_value = "dev")]
-    run_mode: String,
-
-    /// Verbosity level (-v=warning, -vv=info, -vvv=debug).
-    #[clap(short, long, action = clap::ArgAction::Count)]
-    verbose: u8,
-
-    /// Directory where configuration file is located.
-    #[clap(long, default_value = "./config")]
-    pub config_directory: PathBuf,
-}
-
-impl Args {
-    pub async fn execute(&self) -> StdResult<()> {
-        debug!("Run Mode: {}", self.run_mode);
-        let filename = format!("{}/{}.json", self.config_directory.display(), self.run_mode);
-        debug!("Reading configuration file '{}'.", filename);
-        let config: ConfigBuilder<DefaultState> = config::Config::builder()
-            .add_source(config::File::with_name(&filename).required(false))
-            .add_source(self.clone());
-
-        self.command.execute(config).await
-    }
-
-    fn log_level(&self) -> Level {
-        match self.verbose {
-            0 => Level::Error,
-            1 => Level::Warning,
-            2 => Level::Info,
-            3 => Level::Debug,
-            _ => Level::Trace,
-        }
-    }
-
-    fn build_logger(&self) -> Logger {
-        let decorator = slog_term::TermDecorator::new().build();
-        let drain = slog_term::CompactFormat::new(decorator).build().fuse();
-        let drain = slog::LevelFilter::new(drain, self.log_level()).fuse();
-        let drain = slog_async::Async::new(drain).build().fuse();
-
-        Logger::root(Arc::new(drain), slog::o!())
-    }
-}
-
-impl Source for Args {
-    fn clone_into_box(&self) -> Box<dyn Source + Send + Sync> {
-        Box::new(self.clone())
-    }
-
-    fn collect(&self) -> Result<Map<String, Value>, config::ConfigError> {
-        Ok(Map::new())
-    }
-}
 
 #[tokio::main]
 async fn main() -> StdResult<()> {
